+++ app/src/main/res/layout/activity_main.xml
@@ -37,7 +37,7 @@
         android:layout_margin="@dimen/fab_margin"
         android:src="@mipmap/ic_plus"
         app:backgroundTint="@color/colorPrimary"
-        app:layout_behavior="arslanali.ru.moneytracker.fab.AddFabBehavior"
+        app:layout_behavior="@string/fab_behavior"
         app:useCompatPadding="true" />
 
 </android.support.design.widget.CoordinatorLayout>
